Windows: Enabled docker volume
